requirements.txt
pandas
sqlalchemy
pyyaml
pyodbc
openpyxl
striprtf
PySide6
